requirements.txt
africastalking==1.2.3
anytree==2.8.0
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0.post1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
distro==1.4.0
Django==3.1.1
django-admin-csvexport==1.9
django-cors-headers==3.7.0
django-csv-exports==1.0.5
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
isort==5.4.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
MarkupPy==1.14
mccabe==0.6.1
msgpack==0.6.2
numpy==1.20.2
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.3
pandas==1.2.4
pep517==0.8.2
Pillow==7.2.0
progress==1.5
pylint==2.6.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
PyYAML==5.4.1
requests==2.24.0
retrying==1.3.3
schema==0.7.2
six==1.15.0
sqlparse==0.3.1
tablib==3.0.0
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.31
webencodings==0.5.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
